requirements.txt
# Framework y dependencias básicas
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0

# Base de datos
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.0

# Seguridad
python-dotenv==1.0.0
Werkzeug==3.0.1
bcrypt==4.1.1
PyJWT==2.8.0


# Validación y serialización
marshmallow==3.20.1
python-dateutil==2.8.2

# Caché
redis==5.0.1

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# API tools
requests==2.31.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Utilities
click==8.1.7
